+++ nodejs/src/index.mjs
@@ -1,5 +1,8 @@
 import { createRequire } from 'node:module';
 const require = createRequire(import.meta.url);
 const backloopDev = require('./index.js');
-const httpsOptions = await backloopDev.httpsOptionsPromise();
-export default httpsOptions;
+
+export default await backloopDev.httpsOptionsPromise();
+export const httpsOptions = backloopDev.httpsOptions;
+export const httpsOptionsAsync = backloopDev.httpsOptionsAsync;
+export const httpsOptionsPromise = backloopDev.httpsOptionsPromise;