+++ renew/src/index.js
@@ -3,7 +3,6 @@
  * [BSD-3-Clause](https://github.com/perki/backloop.dev/blob/main/LICENSE)
  */
 const acme = require('acme-client');
-const { read } = require('./files');
 const gandi = require('./gandi');
 const savecert = require('./saveccert');
 const pack = require('./pack');
@@ -11,16 +10,15 @@ const pack = require('./pack');
 const DOMAIN = process.env.BACKLOOP_DOMAIN || 'backloop.dev';
 const IS_PRODUCTION = process.env.IS_PRODUCTION || false; // set to true when going to production
 
-if (! process.env.ACME_ACCOUNT_KEY) {
+if (!process.env.ACME_ACCOUNT_KEY) {
   throw new Error('Missing environement var ACME_ACCOUNT_KEY');
 }
 const ACME_ACCOUNT_KEY = process.env.ACME_ACCOUNT_KEY;
 
-if (! process.env.ACME_ACCOUNT_URL) {
+if (!process.env.ACME_ACCOUNT_URL) {
   throw new Error('Missing environement var ACME_ACCOUNT_URL');
 }
 const ACME_ACCOUNT_URL = process.env.ACME_ACCOUNT_URL;
-
 
 (async () => {
   const client = new acme.Client({
@@ -29,11 +27,10 @@ const ACME_ACCOUNT_URL = process.env.ACME_ACCOUNT_URL;
     accountUrl: ACME_ACCOUNT_URL
   });
 
-  const [certificateKey, csr] = await acme.crypto.createCsr({
-      commonName: '*.' + DOMAIN,
-      // altNames: [DOMAIN]  // was causing double request 
+  const [, csr] = await acme.crypto.createCsr({
+    commonName: '*.' + DOMAIN
+    // altNames: [DOMAIN]  // was causing double request
   });
-
 
   console.log('START');
   const certificate = await client.auto({
@@ -43,8 +40,7 @@ const ACME_ACCOUNT_URL = process.env.ACME_ACCOUNT_URL;
     challengeRemoveFn
   });
 
-
-  savecert(DOMAIN, certificate, certificateKey);
+  savecert(DOMAIN, certificate);
   await pack(DOMAIN);
   console.log('DONE');
 })();
@@ -56,5 +52,5 @@ async function challengeCreateFn (authz, challenge, keyAuthorization) {
 
 async function challengeRemoveFn (authz, challenge, keyAuthorization) {
   console.log('****challengeRemoveFn');
-  //await gandi.update(DOMAIN, '_acme-challenge', ['cleared']);
+  // await gandi.update(DOMAIN, '_acme-challenge', ['cleared']);
 }